+++ eng/vscode/CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.5.13 - Unreleased
 
 ## Added
-- Added `azmcp-keyvault-admin-settings-get` command to retrieve Key Vault administration settings (currently only purge protection state and soft delete retention days).
+- Added `azmcp-keyvault-admin-settings-get` command to retrieve Key Vault account settings.
 
 
 # Release History